rydder litt i kode og sjekker funksjonalitet
